requirements.txt
backcall==0.2.0
basemap @ https://github.com/matplotlib/basemap/archive/master.zip
certifi==2020.6.20
click==8.0.3
decorator==4.4.2
Flask==2.0.2
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596207638929/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1598883821407/work
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi @ file:///tmp/build/80754af9/jedi_1611333125342/work
Jinja2==3.0.3
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.3
OWSLib==0.25.0
pandas==1.3.4
parso @ file:///tmp/build/80754af9/parso_1596826841367/work
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
ptyprocess==0.6.0
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
pyproj==3.2.1
pyshp==2.1.3
python-dateutil==2.8.1
pytz==2021.3
PyYAML==6.0
pyzmq==19.0.2
six==1.15.0
torch==1.10.0
torchvision==0.11.1
tornado==6.0.4
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
typing-extensions==3.10.0.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
Werkzeug==2.0.2
